Revamp repository
It has been four years since I've changed anything in this library, and
this commit will make the code more up-to-date.

- Configure GitHub actions
- Configure integration with Codecov
- Configure PHPStan
- Remove integration with Scrutinizer
- Remove integration with Travis
- Upgrade Composer scripts
- Upgrade PHPUnit's version
- Upgrade Respect's coding standard
- Upgrade LICENSE file
- Upgrade license headers

Because of the new coding standards, there have been some backward
compatibility changes, but that's not a problem because it won't change
much of the final API.

Because of the upgraded packages, I had to drop the support PHP versions
before 8.1.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>