Change how to stringify JsonSerializable objects
We can only convert those objects into JSON via the `jsonSerialize()`
method. Being more descriptive that the JSON we see is the result of
that method will make it more straightforward.

Besides that, it's better to directly use a stringifier that converts
data to JSON instead of calling the whole stringifier cluster to
stringify the result of `jsonSerialize()`.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>