Change how to stringify Stringable objects
This library's goal is to convert anything to a string, and although it
seems obvious to return the content of the `__toString()` method, that
behavior does not tell the user much about what the input is.

This commit sets the tone for this library: it's for engineers to
represent anything as a string. However, the fact that we can convert
an object to a string does not mean it is a string.

This commit will also change how we verify those objects. Since PHP 8.0,
any object that implements a `__toString()` method automatically
implements the Stringable interface. Since we dropped support for
versions that are less than 8.1, it makes sense to check that instead of
checking if the method exists.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>