Change how to stringify Iterator objects
I'm changing the notation to ensure we keep consistency with the recent
changes, so I'm using "{}" to represent the body and "current()" to make
it explicit that it's a method call.

I changed the behavior of this stringifier. After all, it might be
unwanted to iterate over the whole object to stringify it because it
might do an expensive operation.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>