Change how to stringify throwable objects
I'm matching the recent format changes, using "{}" to represent objects.
In the case of throwables, I thought it would make more sense to have
the message and the location, which is the exact behavior of the first
line of the output of the `__toString()` method of an exception. That
means we no longer include the exception code.

Also, for optimization, I'm changing the stringifier that the
`ThrowableObjectStringifier` uses because I know it only stringifies a
string, hence using the `JsonEncodableStringifier` improves performance
as we don't need to call all other stringifies.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>