Change how to stringify objects
I'm matching the latest changes in format, using "{}" for objects, but
I'm making many improvements to the `ObjectStringifier`.

When stringifying objects, we have access to the protected and private
properties. I made it more familiar as well, prefixing the variables
with a "$", which is how we describe variables, and using the symbols
"+", "#" and "-", which is how UML represents visibility.

I also fixed a bug because getting the value of non-initialized
properties raises an error, which we didn't consider before.

With this commit, I also introduced a maximum number of properties to
prevent longer strings and a maximum depth to infinite recursion.

Since now we always represent objects with "{}", we don't need to prefix
it with "object" anymore.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>