Change how to stringify anonymous classes
The class name of an anonymous class can be long, so I decided to make
it shorter. A simple "class" is enough to describe the class name since
it is a reserved word, and that's how we define an anonymous class.

If the anonymous class extends another class, it makes sense to keep the
extended class name while ensuring the user understands it is anonymous.
In those cases, I kept the same behavior as PHP, with "@anonymous" next
to the extended class name.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>