Change how to stringify objects with __debugInfo()
That method intends to enable engineers to have a better overview of the
object when they want to debug it. That said, using that method when
it's present is better.

The representation is similar to the one I used on Stringeable objects,
in which we have clear that we're executing a specific method rather
than just showing the values.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>