Improve the reusability of the concrete API
Even though this library is extensible, I acknowledge that combining a
custom stringifier with the existing ones takes work.

I made the API a bit more reusable and created a `Stringify` class that
could facilitate engineers to use this library from a DIC.

There were also a couple of name changes because "cluster" always
bothered me a little.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>